Refactor Di appComponent, rename mainFragment to ImageList
