+++ app/src/main/java/jp/kiriuru/pixabaytest/di/ViewModelBuilderModule.kt
@@ -1,0 +1,14 @@
+package jp.kiriuru.pixabaytest.di
+
+import androidx.lifecycle.ViewModelProvider
+import dagger.Binds
+import dagger.Module
+import jp.kiriuru.pixabaytest.ui.imageList.MainViewModelFactory
+
+@Module
+interface ViewModelBuilderModule {
+    @Binds
+    fun bindViewModelFactory(
+        factory: MainViewModelFactory
+    ): ViewModelProvider.Factory
+}
