+++ app/src/main/java/jp/kiriuru/pixabaytest/ui/detail/DetailFragment.kt
@@ -4,9 +4,7 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.activity.OnBackPressedCallback
 import androidx.fragment.app.Fragment
-import androidx.navigation.fragment.findNavController
 import com.bumptech.glide.load.engine.DiskCacheStrategy
 import jp.kiriuru.pixabaytest.data.model.Hits
 import jp.kiriuru.pixabaytest.databinding.DetailFragmentBinding
@@ -14,7 +12,6 @@ import jp.kiriuru.pixabaytest.utils.Const.Companion.BUNDLE
 import jp.kiriuru.pixabaytest.utils.GlideApp
 
 class DetailFragment : Fragment() {
-    companion object;
 
     private var _binding: DetailFragmentBinding? = null
     private val binding get() = checkNotNull(_binding)
@@ -38,7 +35,6 @@ class DetailFragment : Fragment() {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        onBackPressed()
 
         GlideApp.with(view.context).load(hits.largeImageURL)
             .diskCacheStrategy(DiskCacheStrategy.ALL)
@@ -53,21 +49,5 @@ class DetailFragment : Fragment() {
         super.onDestroy()
         _binding = null
     }
-
-    private fun onBackPressed() {
-        activity?.onBackPressedDispatcher?.addCallback(
-            viewLifecycleOwner,
-            object : OnBackPressedCallback(true) {
-                override fun handleOnBackPressed() {
-                    val isNavigatedUp = findNavController().navigateUp()
-                    if (isNavigatedUp) {
-                        return
-                    } else {
-                        activity?.finish()
-                    }
-                }
-            }
-        )
-    }
-
 }
+
