+++ app/src/main/java/jp/kiriuru/pixabaytest/ui/imageList/ImageListFragment.kt
@@ -1,4 +1,4 @@
-package jp.kiriuru.pixabaytest.ui.main
+package jp.kiriuru.pixabaytest.ui.imageList
 
 import android.content.Context
 import android.os.Bundle
@@ -20,10 +20,10 @@ import androidx.navigation.fragment.findNavController
 import androidx.recyclerview.widget.GridLayoutManager
 import jp.kiriuru.pixabaytest.App
 import jp.kiriuru.pixabaytest.R
-import jp.kiriuru.pixabaytest.data.adapter.RvAdapter
+import jp.kiriuru.pixabaytest.data.adapter.ImageListAdapter
 import jp.kiriuru.pixabaytest.data.adapter.decoration.ImageDecoration
 import jp.kiriuru.pixabaytest.data.model.Hits
-import jp.kiriuru.pixabaytest.databinding.FragmentMainBinding
+import jp.kiriuru.pixabaytest.databinding.FragmentListImageBinding
 import jp.kiriuru.pixabaytest.utils.ClickListener
 import jp.kiriuru.pixabaytest.utils.Const.Companion.BUNDLE
 import jp.kiriuru.pixabaytest.utils.Const.Companion.TAG_MAIN
@@ -33,7 +33,7 @@ import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 
-class MainFragment : Fragment(), ClickListener<Hits> {
+class ImageListFragment : Fragment(), ClickListener<Hits> {
 
     private var defaultPerImage: Int = 30
     private var defaultSearchReq: String = ""
@@ -41,17 +41,18 @@ class MainFragment : Fragment(), ClickListener<Hits> {
     @Inject
     lateinit var viewModelFactory: ViewModelProvider.Factory
 
-    private val viewModel by viewModels<MainViewModel> { viewModelFactory }
+    private val viewModel by viewModels<ImageListViewModel> { viewModelFactory }
 
-    private var _binding: FragmentMainBinding? = null
+    private var _binding: FragmentListImageBinding? = null
     private val binding get() = checkNotNull(_binding)
-    private lateinit var mAdapter: RvAdapter
+
+    private lateinit var mAdapter: ImageListAdapter
 
 
     override fun onAttach(context: Context) {
         super.onAttach(context)
 
-        (requireActivity().application as App).appComponent.mainComponent()
+        (requireActivity().application as App).appComponent.imageListComponent()
             .create().inject(this)
     }
 
@@ -59,12 +60,13 @@ class MainFragment : Fragment(), ClickListener<Hits> {
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        _binding = FragmentMainBinding.inflate(layoutInflater)
+        _binding = FragmentListImageBinding.inflate(layoutInflater)
         return binding.root
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
+
 
         initRV()
 
@@ -112,7 +114,7 @@ class MainFragment : Fragment(), ClickListener<Hits> {
     }
 
     private fun initRV() {
-        mAdapter = RvAdapter(this)
+        mAdapter = ImageListAdapter(this)
 
         with(binding.recycleView) {
             layoutManager = GridLayoutManager(context, 2)
