+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -7,9 +7,9 @@
 
     <fragment
         android:id="@+id/image_list"
-        android:name="jp.kiriuru.pixabaytest.ui.main.MainFragment"
+        android:name="jp.kiriuru.pixabaytest.ui.imageList.ImageListFragment"
         android:label="Search Images"
-        tools:layout="@layout/fragment_main">
+        tools:layout="@layout/fragment_list_image">
         <action
             android:id="@+id/action_image_list_to_image_detail"
             app:destination="@id/image_detail" />
